+++ wgpu-core/src/timestamp_normalization/mod.rs
@@ -253,8 +253,16 @@ impl TimestampNormalizer {
 
     /// Create a bind group for normalizing timestamps in `buffer`.
     ///
-    /// This function is unsafe because it does not know that `buffer_size` is
-    /// the true size of the buffer.
+    /// `buffer_size` must be 4B-aligned, so that it is valid as a storage binding,
+    /// must be the same or more than the application-visible buffer size, and must
+    /// not overrun the buffer. This function is unsafe because it does not know and
+    /// cannot check whether `buffer_size` meets these requirements.
+    ///
+    /// The buffer size passed here is used to construct a binding that must be
+    /// able to access the full application-visible range of the buffer. It is
+    /// not used to check the destination range of query resolution requests. That
+    /// uses the application-requested buffer size, obtained from the `wgpu_core`
+    /// `Buffer`.
     pub unsafe fn create_normalization_bind_group(
         &self,
         device: &Device,