+++ wgpu/src/api/render_pass.rs
@@ -109,8 +109,7 @@ impl RenderPass<'_> {
             &buffer_slice.buffer.inner,
             index_format,
             buffer_slice.offset,
-            // TODO(https://github.com/gfx-rs/wgpu/issues/3170): Empty slices should be supported here
-            Some(buffer_slice.size_expect_nonzero()),
+            Some(buffer_slice.size),
         );
     }
 
@@ -140,8 +139,7 @@ impl RenderPass<'_> {
                 slot,
                 Some(&buffer_slice.buffer.inner),
                 buffer_slice.offset,
-                // TODO(https://github.com/gfx-rs/wgpu/issues/3170): Empty slices should be supported here
-                Some(buffer_slice.size_expect_nonzero()),
+                Some(buffer_slice.size),
             );
         } else {
             self.inner.set_vertex_buffer(slot, None, 0, None);